Re-work as reagent
Ugh, clearly hooks aren't meant to be used from Clojurescript. It also
seems like they don't have to: raw React and react-table has a lot of
concerns around memoization and equality comparison that we don't share
in cljs. Trying out an approach of maintaining a react-table inspired
API but keeping the table state in an ratom.