Add submit button to input component
Adding this makes the form submission "take" on a return in the input
field. Doesn't matter how many we have, but if we *don't* have one the
keyDown doesn't propagate up to cause the form submission. (We'll have
an explicit "save" icon thing that drops input focus anyhow, for users
for whom that's not obvious, provides a tab target, etc).

Added some CSS animation and useEffect hijinks to draw the user's eye to
values that have changed at the moment they change in the UI.